create nested_training_parameters.py
