+++ GANDLF/Configuration/user_defined_parameters.py
@@ -2,6 +2,7 @@ from typing import Union
 from pydantic import BaseModel, model_validator, Field, AfterValidator, ConfigDict
 
 from GANDLF.Configuration.default_parameters import DefaultParameters
+from GANDLF.Configuration.nested_training_parameters import NestedTraining
 from GANDLF.Configuration.scheduler_parameters import Scheduler
 from GANDLF.config_manager import version_check
 from importlib.metadata import version
@@ -18,32 +19,6 @@ class Version(BaseModel):  # TODO: Maybe should be to another folder
     def validate_version(self) -> Self:
         if version_check(self.model_dump(), version_to_check=version("GANDLF")):
             return self
-
-
-class NestedTraining(BaseModel):  # TODO: Maybe should be in another folder
-    stratified: bool = Field(
-        default=False,
-        description="this will perform stratified k-fold cross-validation but only with offline data splitting",
-    )
-    testing: int = Field(
-        default=-5,
-        description="this controls the number of testing data folds for final model evaluation; [NOT recommended] to disable this, use '1'",
-    )
-    validation: int = Field(
-        default=-5,
-        description="this controls the number of validation data folds to be used for model *selection* during training (not used for back-propagation)",
-    )
-    proportional: bool = Field(default=None)
-
-    @model_validator(mode="after")
-    def validate_nested_training(self) -> Self:
-        if self.proportional is not None:
-            self.stratified = self.proportional
-        return self
-
-
-
-
 
 class UserDefinedParameters(DefaultParameters):
     version: Version = Field(
